requirements.txt
Pillow>=11.1.0
pillow-heif>=1.1.0
pygame>=2.6.1
pyinstaller>=6.12.0
